backend/requirements.txt
flask
python-dotenv
pyjwt
flasgger
werkzeug
requests
yfinance
flask-cors
boto3
uuid
botocore
moto[boto3]
pytest